+++ tarantool/src/network/client/reconnect.rs
@@ -157,6 +157,7 @@ mod tests {
             listen_port(),
             protocol::Config {
                 creds: Some(("test_user".into(), "password".into())),
+                ..Default::default()
             },
         )
     }